Finalize changes
